Clarify that we mean strategic members of the WG
For votes to approve the designation of a profile or platform, we require a super-majority of the strategic members of the work group. With this patch, we disambiguate "strategic members" to mean specifically strategic members of the working group. We depend on the working group's charter to define "strategic member".