Specify a minimum of one week for votes.
Per Issue #3, this patch adds that specification committee votes must be scheduled to run for a period of no less that one week.

This patch also clarifies that we need a "Super-majority **vote**". It also includes a minor tweak to the bullet list grammar by substituting semi-colons for periods on each bullet item.