Use the term "ballot"
This patch addresses Issue #8 and Issue #9. Rather than refer to the
process as a "vote", we use the term "ballot" (individual members cast
their vote). We also address the case where a simple typo (or otherwise
minor change that does not change semantic intent) is required during an
ongoing ballot.